The code calculates the total cost of a ticket
The objective of this code is to calculate the total cost of a ticket based on the number of kilometers to be traveled. The user inputs the number of kilometers via standard input (console), then the program calculates the total cost of the ticket using an instance of the `CalculoKilometros` class. Finally, the total cost is displayed in the console in Colombian peso currency format, using comma as the thousands separator and the currency symbol, with two decimal places. Once the operation is completed, the Scanner object used for input is closed.