StudentDataCollector
This code appears to be similar to the previous one, but it uses an `InputHelper` class to facilitate input from the user. It still collects information about a group of students, including their ages and heights, and then calculates and displays various statistics about the group, such as the average age and height, as well as the number of students over 18 years old and those taller than 1.75 meters.